GitHub Actions: implement parallel testing
This allows for parallel execution of the echidna tests, which
helps improve testing speed, especially once newer solc versions
are included in the pipeline.

+++ .github/scripts/install-solc.sh
@@ -1,12 +1,8 @@
-#! /bin/bash
+#!/bin/bash
 
 # Adapted from https://github.com/commercialhaskell/stack
 
 set -eux
-
-if [ -f $HOME/.local/bin/solc-0.4.25 ] && [ -f $HOME/.local/bin/solc-0.5.7 ]; then
-  exit 0
-fi
 
 mkdir -p $HOME/.local/bin;
 
@@ -16,15 +12,29 @@ travis_retry() {
 }
 
 fetch_solc_linux() {
-  rm -Rf solc-static-linux;
-  wget https://github.com/ethereum/solidity/releases/download/v0.4.25/solc-static-linux;
-  chmod +x solc-static-linux;
-  mv solc-static-linux $HOME/.local/bin/solc-0.4.25;
-  wget https://github.com/ethereum/solidity/releases/download/v0.5.7/solc-static-linux;
-  chmod +x solc-static-linux;
-  mv solc-static-linux $HOME/.local/bin/solc-0.5.7;
+  VER="$1"
+  if [ ! -f "$HOME/.local/bin/solc-$VER" ]; then
+    rm -Rf solc-static-linux
+    wget "https://github.com/ethereum/solidity/releases/download/v$VER/solc-static-linux"
+    chmod +x solc-static-linux
+    mv solc-static-linux "$HOME/.local/bin/solc-$VER"
+    echo "Downloaded solc $VER"
+  else
+    echo "Skipped solc $VER, already present"
+  fi
+}
+
+fetch_all_solc_linux() {
+  fetch_solc_linux "0.4.25"
+  fetch_solc_linux "0.5.7"
+  fetch_solc_linux "0.6.12"
+  fetch_solc_linux "0.7.5"
 }
 
 if [ "$HOST_OS" = "Linux" ]; then
-  travis_retry fetch_solc_linux
+  if [ "${SOLC_VER:-}" == "" ]; then
+    travis_retry fetch_all_solc_linux
+  else
+    travis_retry fetch_solc_linux "$SOLC_VER"
+  fi
 fi
